refactor(bindings): rename BleManager to BleTransportFacade and remove legacy file
Merges the legacy com.offlineprotocol.BleManager into
com.offlineprotocol.ble.nordic.BleTransportFacade via git rename, so
there is no longer a BleManager class in our codebase. Preserves git
blame history through the move.

Concrete changes to the moved file:
  - package: com.offlineprotocol -> com.offlineprotocol.ble.nordic
  - class BleManager -> class BleTransportFacade
  - TAG: "BleManager" -> "BleTransportFacade"
  - this@BleManager -> this@BleTransportFacade (3 sites)
  - Added imports for TransportManager / TransportState /
    TransportManagerListener / TransportException /
    BleDiscoveryBootstrapPolicy — previously resolved via same-package
    access, now require explicit imports since the class moved into
    the ble.nordic subpackage.
  - Removed the former BleTransportFacade stub that composed a legacy
    BleManager delegate; the facade IS the implementation now.

NordicAdvertiser / NordicGattServer references stay unqualified: the
facade is in their same package so same-package resolution covers
them without explicit imports.

Central-role logic (scanning, GATT client callback, fragment
accounting, connection retry, mesh orchestration) still lives inside
this file — the migration to NordicPeerClient is deferred to a
follow-up PR. What this commit accomplishes is the structural end
state of the plan's Step 7: no legacy BleManager class exists, all
BLE code lives under ble/ or ble/nordic/, and OfflineProtocolModule
constructs BleTransportFacade as its BLE transport entry point.
